+++ src/star/star_solo/test.sh
@@ -1,16 +1,13 @@
 #!/bin/bash
 
-set -e
+set -e -o pipefail
 
 ## VIASH START
-meta_executable="target/docker/star/star_solo/star_solo"
-meta_resources_dir="src/star/star_solo"
+meta_executable="target/docker/star/star_align_reads/star_align_reads"
+meta_resources_dir="src/star/star_align_reads"
 ## VIASH END
 
 #########################################################################################
-
-## TODO: add star solo tests
-
 
 # helper functions
 assert_file_exists() {
@@ -92,15 +89,16 @@ cd star_align_reads_se
 echo "> Run star_align_reads on SE"
 "$meta_executable" \
   --input "../reads_R1.fastq" \
-  --genomeDir "../index/" \
+  --genome_dir "../index/" \
   --aligned_reads "output.sam" \
   --log "log.txt" \
-  --outReadsUnmapped "Fastx" \
+  --out_reads_unmapped "Fastx" \
   --unmapped "unmapped.sam" \
-  --quantMode "TranscriptomeSAM;GeneCounts" \
+  --quant_mode "TranscriptomeSAM;GeneCounts" \
   --reads_per_gene "reads_per_gene.tsv" \
-  --outSJtype Standard \
+  --out_sj_type Standard \
   --splice_junctions "splice_junctions.tsv" \
+  --reads_aligned_to_transcriptome "transcriptome_aligned.bam" \
   ${meta_cpus:+---cpus $meta_cpus}
 
 # TODO: Test data doesn't contain any chimeric reads yet
@@ -144,10 +142,10 @@ echo ">> Run star_align_reads on PE"
 "$meta_executable" \
   --input ../reads_R1.fastq \
   --input_r2 ../reads_R2.fastq \
-  --genomeDir ../index/ \
+  --genome_dir ../index/ \
   --aligned_reads output.bam \
   --log log.txt \
-  --outReadsUnmapped Fastx \
+  --out_reads_unmapped Fastx \
   --unmapped unmapped_r1.bam \
   --unmapped_r2 unmapped_r2.bam \
   ${meta_cpus:+---cpus $meta_cpus}